+++ curta/src/chip/ec/edwards/scalar_mul/generator.rs
@@ -14,7 +14,6 @@ use serde::{Deserialize, Serialize};
 
 use super::air::ScalarMulEd25519;
 use super::gadget::EdScalarMulGadget;
-use crate::chip::builder::AirTraceData;
 use crate::chip::ec::edwards::ed25519::{Ed25519, Ed25519BaseField, Ed25519Parameters};
 use crate::chip::ec::gadget::EllipticCurveWriter;
 use crate::chip::ec::point::{AffinePoint, AffinePointRegister};
@@ -25,6 +24,7 @@ use crate::chip::register::array::ArrayRegister;
 use crate::chip::register::element::ElementRegister;
 use crate::chip::register::RegisterSerializable;
 use crate::chip::trace::generator::ArithmeticGenerator;
+use crate::chip::trace::trae_data::AirTraceData;
 use crate::chip::utils::{biguint_to_16_digits_field, biguint_to_bits_le, field_limbs_to_biguint};
 use crate::chip::Chip;
 use crate::math::extension::CubicParameters;
